HttpConnectionBuilderOptions
